# Avoid millions of rand() calls() when running tests