--- a/libarchive/test/test_read_large.c
+++ b/libarchive/test/test_read_large.c
@@ -23,6 +23,7 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 #include "test.h"
+#include "test_utils.h"
 __FBSDID("$FreeBSD: head/lib/libarchive/test/test_read_large.c 201247 2009-12-30 05:59:21Z kientzle $");
 
 static unsigned char testdata[10 * 1024 * 1024];
@@ -37,7 +38,6 @@ static unsigned char buff[11 * 1024 * 1024];
 /* Check correct behavior on large reads. */
 DEFINE_TEST(test_read_large)
 {
-	unsigned int i;
 	int tmpfilefd;
 	char tmpfilename[] = "test-read_large.XXXXXX";
 	size_t used;
@@ -45,8 +45,7 @@ DEFINE_TEST(test_read_large)
 	struct archive_entry *entry;
 	FILE *f;
 
-	for (i = 0; i < sizeof(testdata); i++)
-		testdata[i] = (unsigned char)(rand());
+	fill_with_pseudorandom_data(testdata, sizeof(testdata));
 
 	assert(NULL != (a = archive_write_new()));
 	assertA(0 == archive_write_set_format_ustar(a));