--- a/libarchive/test/test_write_format_7zip_large.c
+++ b/libarchive/test/test_write_format_7zip_large.c
@@ -25,6 +25,7 @@
 
 
 #include "test.h"
+#include "test_utils.h"
 __FBSDID("$FreeBSD$");
 
 #define LARGE_SIZE	(16*1024*1024)
@@ -37,7 +38,6 @@ test_large(const char *compression_type)
 	size_t buffsize = LARGE_SIZE + 1024 * 256;
 	size_t datasize = LARGE_SIZE;
 	char *buff, *filedata, *filedata2;
-	unsigned i;
 
 	assert((buff = malloc(buffsize)) != NULL);
 	assert((filedata = malloc(datasize)) != NULL);
@@ -87,8 +87,7 @@ test_large(const char *compression_type)
 		/* NOTE: PPMd cannot handle random data correctly.*/
 		memset(filedata, 'a', datasize);
 	} else {
-		for (i = 0; i < datasize; i++)
-			filedata[i] = (char)rand();
+		fill_with_pseudorandom_data(filedata, datasize);
 	}
 	assertEqualInt(datasize, archive_write_data(a, filedata, datasize));