--- a/test_utils/test_utils.h
+++ b/test_utils/test_utils.h
@@ -27,6 +27,9 @@
 #ifndef TEST_UTILS_H
 #define TEST_UTILS_H
 
+#include <stddef.h>
+#include <stdint.h>
+
 struct test_list_t
 {
   void (*func)(void);
@@ -35,5 +38,7 @@ struct test_list_t
 };
 
 int get_test_set(int *, int, const char *, struct test_list_t *);
+void fill_with_pseudorandom_data(void* buffer, size_t size);
+void fill_with_pseudorandom_data_seed(uint64_t seed, void* buffer, size_t size);
 
 #endif /* TEST_UTILS_H */