@jsonn 2020-09-17
I don't mind refactoring the loops into a common helper, but I don't like dropping rand here. Just using all 32bit of the return value (and not just 8bit) should already cover much of the performance difference.