@jsonn 2020-09-17
It's a poor man's fuzzing device, so yeah.