@arichardson 2020-10-19
This new version uses rand() to seed the xorshift64 RNG, so the per-platform different values should still exist.

The performance improvement is not as big on native x86_64 hardware, but test_write_format_7zip_large_copy still goes from 460ms to 380ms (i.e. about 20% faster).