@arichardson 2020-10-19
> This new version uses rand() to seed the xorshift64 RNG, so the per-platform different values should still exist.
> 
> The performance improvement is not as big on native x86_64 hardware, but test_write_format_7zip_large_copy still goes from 460ms to 380ms (i.e. about 20% faster).

Sorry thos numbers were from a debug build, on a release build it's much more noticeable: 314.9 ms ±   3.9 ms before, 227.8 ms ±   5.8 ms after (i.e. 38% faster for that test)