@arichardson 2020-11-30
@jsonn Does this version look better? It should have the same features as the rand() based version while being faster.

It makes a quite significant difference for our QEMU CI jobs, and I'd quite like to avoid carrying downstream patches.