@mmatuska 2020-12-01
@jsonn @kientzle we should also consider disabling fuzzing tests. oss-fuzz is doing a much better job for us there.