@kientzle 2022-01-12
@mmatuska Yeah, I agree that oss-fuzz is doing a lot more for us.  But having a fast set of basic fuzzing tests here does make me feel a little more confident, so if folks want to spend some time making those better, faster, and/or more random, I'm all in favor.