@arichardson 2022-01-15
rebased and made the helper static.